requirements.txt
flask>=2.3.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
schedule>=1.2.0
